Cleanup

+++ src/main/java/net/tb11/LostHorizons/crops/ArcaniteCrystal.java
@@ -32,7 +32,7 @@ public class ArcaniteCrystal extends CropBlock{
         return LostHorizons.Arcanite_Seed;
     }
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return Growth_Stage[(Integer)state.get(this.getAgeProperty())];
+        return Growth_Stage[state.get(this.getAgeProperty())];
     }
 
     @Override
