+++ src/main/java/net/tb11/LostHorizons/crops/CrylatiteCrystal.java
@@ -35,7 +35,7 @@ public class CrylatiteCrystal extends CropBlock{
         return LostHorizons.Crylatite_Seed;
     }
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return Growth_Stage[(Integer)state.get(this.getAgeProperty())];
+        return Growth_Stage[state.get(this.getAgeProperty())];
     }
     
     @Override
