+++ src/main/java/net/tb11/LostHorizons/crops/NoxiteCrystal.java
@@ -35,7 +35,7 @@ public class NoxiteCrystal extends CropBlock{
         return LostHorizons.Noxite_Seed;
     }
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return Growth_Stage[(Integer)state.get(this.getAgeProperty())];
+        return Growth_Stage[state.get(this.getAgeProperty())];
     }
     
     @Override
